+++ source/extensions/omni.isaac.kit/omni/isaac/kit/simulation_app.py
@@ -379,6 +379,9 @@ class SimulationApp:
 
     def close(self) -> None:
         """Close the running Omniverse Toolkit."""
+        # workaround for exit issues, clean the stage first:
+        omni.usd.get_context().new_stage()
+        omni.kit.app.get_app().update()
         # check if exited already
         if not self._exiting:
             self._exiting = True