+++ source/extensions/omni.isaac.range_sensor/docs/CHANGELOG.md
@@ -2,6 +2,12 @@
 CHANGELOG
 **********
 
+[0.2.2] - 2022-03-10
+========================
+
+Changed
+========
+- Use orient op for commands that create sensors
 
 
 [0.2.1] - 2022-03-04